requirements.txt
flask
google-genai
markdown
gunicorn
python-dotenv
